+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Current values are:
    * `ALL` - Enables all of the above
 (May still require changes to your logging configuration to see the new logs.)
 * Enables skipping the bundled lib by setting the system property `com.amazon.corretto.crypto.provider.useExternalLib` [PR #168](https://github.com/corretto/amazon-corretto-crypto-provider/pull/168)
+* External integration tests now skip certificate validation for expired certificates.
+   This is to work around external sites which may have allowed their certificates to expire.
+   [PR #190](https://github.com/corretto/amazon-corretto-crypto-provider/pull/189)
 
 ### Patches
 * Improve zeroization of DRBG output. [PR #162](https://github.com/corretto/amazon-corretto-crypto-provider/pull/162)